% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{arr2df}
\alias{arr2df}
\title{extract data from array into a dataframe for selected locations and substances,}
\usage{
arr2df(arr, locmod, submod)
}
\arguments{
\item{arr}{the array to be extracted.}

\item{locmod}{the locations in the array to be extracted}

\item{submod}{the substances in the array to be extracted}
}
\value{
A dataframe with model output values for \code{submod} and \code{locmod}.
}
\description{
extract data from array into a dataframe for selected locations and substances,
}
\examples{
arr <- his2arr(example.his)
df.mod <- arr2df(arr, "location", "substance")
str(df.mod)
}

