% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfunctions.R
\name{circleFun}
\alias{circleFun}
\title{produce dataframe with circle points for plotting}
\usage{
circleFun(center = c(0, 0), diameter = 1, npoints = 100)
}
\value{
A dataframe with npoints points located on a circle
circleFun()
}
\description{
produce dataframe with circle points for plotting
}

