% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotfunctions.r
\name{DelwaqEcoplot}
\alias{DelwaqEcoplot}
\title{plot time series variable with limiting factors}
\usage{
DelwaqEcoplot(arr, locmod, submod, limmod, plottype)
}
\arguments{
\item{arr}{array with model results produced with \code{his2arr}}

\item{locmod}{one or more locations}

\item{submod}{one or more variables}

\item{limmod}{limiting factors to plot}

\item{plottype}{1-solid lines of varying thickness; 2- thick lines of varying transparancy}
}
\value{
A timeseries plot with variable and limiting factors
}
\description{
plot time series variable with limiting factors
}

