% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotfunctions.r
\name{DelwaqEcoplot2}
\alias{DelwaqEcoplot2}
\title{plot time series variable with limiting factors}
\usage{
DelwaqEcoplot2(arr, locmod, submod, limmod, plottype)
}
\arguments{
\item{arr}{array with model results produced with \code{his2arr}}

\item{locmod}{one or more locations}

\item{submod}{one or more variables}

\item{limmod}{limiting factors to plot}

\item{plottype}{options: 1-solid lines of varying thickness; 2- thick lines of varying transparancy}
}
\value{
A timeseries plot with variable and limiting factors
}
\description{
plot time series variable with limiting factors
}
\examples{
library(DelwaqR)
arr <- his2arr(filename = "DATA/NZBLOOM.his", timestamp = F, begintime = "2003-01-01 00:00:00")
dimnames(arr)
submod <- c("Chlfa", "OXY")
locmod <- c("NZR6NW020", "NZR9TS010")
df <- arr2df(arr, locmod=locmod, submod=submod)
df$value[df$variable == "fResptot"] <- -df$value[df$variable == "fResptot"]
library(ggplot2)
plot <- ggplot(df, aes(time, value))
plot +
  geom_line(aes(color = variable), size = 1) +
  geom_point(aes(color = variable), fill = "white",  shape = 21, size = 4) +
  facet_grid((. ~ location))
limmod = c("Limit e", "Limit nit", "Limit pho", "Limit sil")
DelwaqEcoplot(arr = arr, locmod = locmod, submod = submod, limmod = limmod, plottype = 1)
DelwaqEcoplot2(arr = arr, locmod = locmod, submod = submod, limmod = limmod, plottype = 1)
}

