% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{DF_MWTL_NCDF}
\alias{DF_MWTL_NCDF}
\title{Read MWTL NetCDF and convert it to a Dataframe}
\usage{
DF_MWTL_NCDF(substance_mwtl, list_locations_mwtl, workdir)
}
\arguments{
\item{substance_mwtl}{substances with naming as in mwtl ncdf files}

\item{list_locations_mwtl}{locations with naming as in mwtl ncdf files}

\item{workdir}{directory where downloaded ncdf files are stored}
}
\value{
dataframe with available mwtl measurements
}
\description{
Read MWTL NetCDF and convert it to a Dataframe
}

