% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{his2arr}
\alias{his2arr}
\title{convert Delwaq his file into an R array object,}
\usage{
his2arr(filename)
}
\arguments{
\item{filename}{the .his file to be converted.}
}
\value{
An R array object of the Delwaq .his file named  \code{filename}.
}
\description{
convert Delwaq his file into an R array object,
}
\examples{
arr <- his2arr(example.his)
dimnames(arr)
}

