% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statfunctions.R
\name{make.target.table3}
\alias{make.target.table3}
\title{extract data from array into a dataframe for selected locations and substances,}
\usage{
make.target.table3(formulax, df, val_obs, val_mod, logtrans = F)
}
\arguments{
\item{formulax}{formula to define the variables for statistics}

\item{df}{dataframe with observed and modelled values}

\item{val_obs}{the observed variable name}

\item{val_mod}{the modelled variable name}

\item{logtrans(logical)}{whether to logtransform the data before analysis}
}
\value{
A dataframe with statistics to plot target diagram
}
\description{
extract data from array into a dataframe for selected locations and substances,
}

