% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotfunctions.r
\name{plotstacked}
\alias{plotstacked}
\title{plot time series of stacked variables}
\usage{
plotstacked(unit, locmod, submod, arr)
}
\arguments{
\item{unit}{the unit to display on the axis}

\item{locmod}{one or more locations}

\item{submod}{one or more substances}

\item{arr}{array with model results produced with \code{his2arr}}
}
\value{
A standard plot of stacked variables
}
\description{
plot time series of stacked variables
}

