% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ReadCSVMWTLFormat}
\alias{ReadCSVMWTLFormat}
\title{Check and Read CSV data}
\usage{
ReadCSVMWTLFormat(csv_file, name_file)
}
\arguments{
\item{csv_file}{file with measurements}

\item{name_file}{file with names?}
}
\value{
dataframe with available measurements available in the csv file
}
\description{
Check and Read CSV data
}

