% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{arr2df}
\alias{arr2df}
\title{extract data from array into a dataframe for selected locations and substances,}
\usage{
arr2df(arr, locmod, submod)
}
\arguments{
\item{arr}{the array to be extracted.}

\item{locmod}{the locations in the array to be extracted}

\item{submod}{the substances in the array to be extracted}
}
\value{
A dataframe with model output values for \code{submod} and \code{locmod}.
}
\description{
extract data from array into a dataframe for selected locations and substances,
}
\examples{
library(DelwaqR)
arr <- his2arr(filename = "extdata/NZBLOOM.his", timestamp = F, begintime = "2003-01-01 00:00:00")
dimnames(arr)
submod <- c("Chlfa", "OXY")
locmod <- c("NZR6NW020", "NZR9TS010")
df <- arr2df(arr, locmod=locmod, submod=submod)
df$value[df$variable == "fResptot"] <- -df$value[df$variable == "fResptot"]
library(ggplot2)
plot <- ggplot(df, aes(time, value))
plot +
  geom_line(aes(color = variable), size = 1) +
  geom_point(aes(color = variable), fill = "white",  shape = 21, size = 4) +
  facet_grid((. ~ location))
}

