% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{his2arr}
\alias{his2arr}
\title{convert Delwaq his file into an R array object,}
\usage{
his2arr(filename, timestamp = T, begintime = "1900-01-01 00:00:00")
}
\arguments{
\item{filename}{the .his file to be converted.}
}
\value{
An R array object of the Delwaq .his file named  \code{filename}.
}
\description{
convert Delwaq his file into an R array object,
}
\examples{
library(DelwaqR)
arr <- his2arr(filename = "extdata/NZBLOOM.his", timestamp = F, begintime = "2003-01-01 00:00:00")
dimnames(arr)
submod <- c("Chlfa", "OXY")
locmod <- c("NZR6NW020", "NZR9TS010")
df <- arr2df(arr, locmod=locmod, submod=submod)
df$value[df$variable == "fResptot"] <- -df$value[df$variable == "fResptot"]
library(ggplot2)
plot <- ggplot(df, aes(time, value))
plot +
  geom_line(aes(color = variable), size = 1) +
  geom_point(aes(color = variable), fill = "white",  shape = 21, size = 4) +
  facet_grid((. ~ location))
}

