% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfunctions.R
\name{make.target.table3}
\alias{make.target.table3}
\title{produce summary statistics from dataframe with matching measured and modelled values}
\usage{
make.target.table3(formulax, df, val_obs, val_mod, logtrans = F)
}
\arguments{
\item{formulax}{formula to define the variables for statistics}

\item{df}{dataframe with observed and modelled values}

\item{val_obs}{the observed variable name}

\item{val_mod}{the modelled variable name}

\item{logtrans(logical)}{whether to logtransform the data before analysis}
}
\value{
A dataframe with statistics (normalized unbiased RMSD and normalized BIAS) to plot target diagram
}
\description{
produce summary statistics from dataframe with matching measured and modelled values
}
\examples{
library(DelwaqR)
make.target.table3(~ substance + location + season, df.statii, "value.x", "value.y", logtrans = F)
}

