% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfunctions.R
\name{make.target.table3}
\alias{make.target.table3}
\title{extract data from array into a dataframe for selected locations and substances,}
\usage{
make.target.table3(formulax, df, val_obs, val_mod, logtrans = F)
}
\arguments{
\item{formulax}{formula to define the variables for statistics}

\item{df}{dataframe with observed and modelled values}

\item{val_obs}{the observed variable name}

\item{val_mod}{the modelled variable name}

\item{logtrans(logical)}{whether to logtransform the data before analysis}
}
\value{
A dataframe with statistics to plot target diagram
}
\description{
extract data from array into a dataframe for selected locations and substances,
}
\examples{
library(DelwaqR)
arr <- his2arr(filename = "DATA/NZBLOOM.his", timestamp = F, begintime = "2003-01-01 00:00:00")
dimnames(arr)
submod <- c("Chlfa", "OXY")
locmod <- c("NZR6NW020", "NZR9TS010")
df <- arr2df(arr, locmod=locmod, submod=submod)
df$value[df$variable == "fResptot"] <- -df$value[df$variable == "fResptot"]
library(ggplot2)
plot <- ggplot(df, aes(time, value))
plot +
  geom_line(aes(color = variable), size = 1) +
  geom_point(aes(color = variable), fill = "white",  shape = 21, size = 4) +
  facet_grid((. ~ location))
}

